Add runtime deps support for JavaLibrary
Summary:
This is useful for java libraries that depend on other targets
with runtime deps (i.e., CppLibrary/CppBinary).

The specific use case we're interested in supporting is including a `cpp_binary` as a resource into a `java_library`. In dev mode `cpp_binary` is not statically linked, so we can't include the binary as a resource and instead include a bash file which references the dev mode binary on disk. It is not guaranteed the binary will be built though, and we are seeing sporadic issues where our `cpp_binary` does not get included. With support for proper runtime deps we will be able to guarantee the binary is built.

https://fb.workplace.com/groups/askbuck/permalink/2724669350914996/ for more context.

Either this is correct or this feature looked deceptively simple to write, in which case I expect to see quite a few comments on this :)

Reviewed By: asp2insp

shipit-source-id: b33ec2074d