Extract out ProductType determination into new TargetGenerator class
Summary:
ProductType is used in generating a PBXNativeTarget, but is determined in several different places. Much of the target generation also occurs in `generateBinaryTarget` which is called by several methods as the bulk of forming target attributes occurs in this method.

This change consolidates determining the ProductType into one method, with tests to verify the logic.

Additionally it creates TargetGenerator, which will evolve over the next few changes to determine all the target attributes for a target. These attributes will be used to write the PBXNativeTarget.

Reviewed By: williamtwilson

shipit-source-id: 21c6f47ea5