Extract build configuration out of ProjectGenerator (1/n)
Summary:
`.xcconfig` build configuration is currently entwined in Project Generator. This diff extract is out into it's own helper class to allow testing base functionality.

Additionally, more configuration building in project generation can be extracted out to this class. This attempts to scope the work to existing methods vs future diffs that will move blocks of code in `generateBinaryTarget` out.

Note, we remove the `getDefaultDebugBuildConfiguration` which are not the defaults in Xcode. Unless specified, we'd prefer to Xcode utilize the default settings.

Reviewed By: williamtwilson

shipit-source-id: f03972d8b0