Remove indirection for generating targets in ProjectGenerator & simplify codepaths
Summary: As of now, to generate a project, there's a lot of indirection that happens, using caching, maps, etc. In an effort to remove all of the mutable state of ProjectGenerator, this diff kills off all of that stuff, making `createXcodeProject` return an actual object with the data created and cleaning up a bunch of mutable state. The next step from here is to clean up requiredBuildTargets, isProjectGenerated, and a few other remaining stateful items.

Reviewed By: williamtwilson

shipit-source-id: 1119745533