Make requiredBuildTargetsBuilder local to createXcodeProject instead of field
Summary:
This diff begins a series of changes to ProjectGenerator that aim to remove the internal mutability of the class with the expressed goal of having it be immutable, configurable at initialization, and then calling `createXcodeProject` would be a deterministic input / output (aka no modifying fields/ivars/etc).

To start, this moves requiredBuildTargetsBuilder from a field and into a local variable of createXcodeProject(); from there, it is simply passed down the call stack to all sites that need it.

shipit-source-id: 5ff7f98f2c