Fix Mach-O parsing
Summary:
People using OnDemand to build Wilde are hitting file scrubber errors, which seems to be related to:

- loading a 64-bit segment
- assuming the filesize of this segment is within a 32-bit range

This seems like a strange assumption, so I'm fixing that here. Instead, I've swapped the assumption so that the value is long, but needs checking whether it's actually in a 32-bit range.

Reviewed By: philipjameson

shipit-source-id: 5b692c71d2