Making idb install function return the id of the bundle installed
Summary: To run the app tests, we need to install the app and then call the run function, that needs the id of the app that was installed. idb returns this id when the install command finishes, so we now are returning it when the function gets called, instead of returning a boolean.

Reviewed By: williamtwilson

shipit-source-id: 8d3d3b8e86