Retain methods of a synthetic class used for inlining
Summary:
If we have an extension function that uses a lambda, or implements an interface, then the Kotlin compiler will create an inner class to take care of this. If the function is an `inline` function, then the inner class that is created needs to be retained, and also needs to have all its methods retained (or at  least the `invoke` methods and the `init` method - for now retaining everything seems reasonable and probably correct).

Working out which classes have been created for an inline function is not easy. Right now the idea is to retain a list of all of the inline function of the top-level classes. Then when we have an inner class, we search for the corresponding top-level class, and if it has inline functions, then we look to see if the name of the inner class corresponds to one of those inline functions.

It's a bit messy but I don't have a better way to do this right now. I'd be interested to hear your thoughts :)

Reviewed By: jkeljo

shipit-source-id: 660178b872