Update ProjectGeneratorTests to pass a workspace target (Merged Header Maps 2/n)
Summary:
We always pass a workspace target to ProjectGenerator but the majority of tests utilize helper functions that pass an empty workspace target. This is the first of several diffs to start always passing a workspace target, which triggers different code flows (primarily allowing for using a merged header map for public headers).

Unfortunately this breaks some tests which utilize symlinks which are being migrated out but cannot be updated yet because all other required configuration changes for creating a merged header map are not in place. These portions of the test will be reenabled further up the stack and landed together.

Reviewed By: Haud

shipit-source-id: c1f89f6f29