Always pass a workspace target (Merged Header Maps 3/n)
Summary:
1. Remove the optionality of workspace target in ProjectGenerator
2. Update tests to always pass a workspace target

This gets us closer to always forcing header maps for public headers.

Unfortunately this breaks some tests which utilize symlinks which are being migrated out but cannot be updated yet because all other required configuration changes for creating a merged header map are not in place. These portions of the test will be reenabled further up the stack and landed together.

Reviewed By: Haud

shipit-source-id: ac7aa4f752