Provide targetsInRequiredProjects to tests
Summary:
Previously, we were providing an empty set for `targetsInRequiredProjects`, which is used for determining which targets to create a public header map for. This meant that all tests would not create a correct public header map.

This diff provides all the initial nodes and the workspace node as `targetsInRequiredProjects`, which along with the previous changes in the stack, allow for generating the correct headermaps with absolute paths (no symlinks).

This means we can reenable previously commented out portions of some tests, as the header entries for these are now present in the public header map.

Reviewed By: Haud

shipit-source-id: 92ce048149