select() without platform should be an error: attempt 2
Summary:
Previous attempt was reverted because it broke build in one of our repositories.

Now that repository was patched to avoid using selects when platform is not specified.

Additionally, this diff is modified to print target name which fails.

Before this diff the error was:

```
Cannot use select() expression when target platform is not specified
```

without any indication of which target or even which `BUCK` file failed.

The error now is:

```
//:test-library: Cannot use select() expression when target platform is not specified
```

(this is tested in `BuildCommandIntegrationTest.testSelectWithoutTargetPlatform`).

Reviewed By: sbalabanov

shipit-source-id: c423e10d5f