compatible_with rule parameter
Summary:
### Tl;DR

Before this diff we already had two ways to declare rule compatibility with a target platform:

* `target_compatible_with` parameter
* `target_compatible_platforms` parameter

This diff introduces `comparible_with` parameter with intention to deprecate and remove `target_compatible_with` and `target_compatible_platforms`.

## `target_compatible_with` parameter

accepts a list of constraints. So it's not possible to express logic like this: target is compatible with any linux or 64-bit Windows.

However, `select` was accidentally allowed for this attribute, thus in one of our repositories we use this pattern:

```
target_compatible_with = select({
    "config//os:linux": ["config//constraints:any"],
    "config//toolchain:windows-x64_64": ["config//constraints:none"],
}),
```

Which is a too verbose and not friendly to new users.

This parameter allowed us successfully set up new platforms in that new repository, but for widespread adition we want something better.

That's why `target_compatible_platforms` attribute was implemented.

## `target_compatible_platforms` parameter

`target_compatible_platforms` is non-configurable (non-selectable) parameter which accepts a list of platforms. However, it has two drawbacks.

First, this parameter is a bit confusing. We have platform inheritance, but platform inheritance is ignored when `target_compatible_platforms` is checked, only the set of constraints is checked.

Second, and more important, it is inconvenient because different nodes are used in target compatibility check and in select keys. Example:

```
cxx_library(
    name = "thread-utils",
    target_compatible_platforms = [
        # these are platforms
        "config//platforms:linux",
        "config//platforms:windows",
    ],
    deps = select({
        # config settings
        "config//os:linux": ["thread-utils-linux"],
        "config//os:windows": ["thread-utils-windows"],
    }),
)
```

## `compatible_with` parameter, introduced by this diff

`compatible_with` parameter accepts a list of config settings:

```
cxx_library(
    name = "thread-utils",
    compatible_platforms = [
        "config//os:linux",
        "config//os:windows",
    ],
    deps = select({
        "config//os:linux": ["thread-utils-linux"],
        "config//os:windows": ["thread-utils-windows"],
    }),
)
```

Note, `compatible_with` parameter is non-configurable (non-selectable), so it does not allow any negation. So it is not possible to express log like: "compatible with any compiler except MSVC" or "compatible with all OS except Android" or "only without sanitizers". Which is probably a good thing.

* for operating systems and compilers, it's better to explicitly enumerate supported operating systems
* for compilers it's probably better to add a constraint like "gcc-like", and add this constraint to all platforms which use gcc or clang
* for sanitizers it's possible to define a constraint "no-sanitizers" and add it to all platforms without sanitizers (we actually already do it in the repository where we use new platform)

## Name of the parameter

Alternative parameter name could be `target_comparible_configs` or `target_compatible_config_settings`, because this parameter enumerates `config_setting` node references.

`compatible_with` is shorter.

## Bazel

Similar parameters in Bazel are called [`compatible_with`](https://docs.bazel.build/versions/master/be/common-definitions.html#common.compatible_with) and `restricted_to`. Their semantics is quite complicated, but generally they are similar to `target_compatible_platforms`.

### Future steps

* remove `target_compatible_platforms` parameter, which is not used anywhere (and migrate `target_compatible_platforms` tests to `compatible_with` not to lose important test cases)
* migrate `target_compatible_with` to `compatible_with` and remove `target_compatible_with` parameter support from Buck

Reviewed By: sbalabanov

shipit-source-id: 57df57713b