Changing the way to pick simulators for testing with idb
Summary: Some people do not have the default simulator in their macs and this was causing idb to not execute the tests because it was not finding the udid of the default device. We now changed the way we choose the device by first picking any booted devices that are available. If none are booted, the method just picks the first one it finds.

Reviewed By: zeyadsalloum

shipit-source-id: a57a6f190f