Add GenruleBuildable en route to Genrule MBR port
Summary:
As the first step in porting all Genrules to MBR, this commit adds GenruleBuildable,
extracted from the step building logic from within the Genrule class. Note that this
commit does not port Genrule to MBR yet, so this commit provides no observable
behavior change (yet).

There is a lot of code in this diff, but an overwhelming majority of it was ported
from Genrule. Genrule does three major things:
	1) Make a src and tmp directory, symlink the sources into the src directory
	2) Create the shell step (with an optional sandbox) for the command
	3) Set up the environment for the shell step (including optional Android tools)

Steps 1 and 2 are relatively straightforward, except that 2 relies on a
SandboxExecutionStrategy to create a ProgramRunner. To do this, this commit puts
the onus of constructing a SandboxProperties on the caller (i.e. Genrule) and
serializes both SandboxExecutionStrategy and SandboxProperties to no-op classes
(NoSandboxExecutionStrategy and an empty SandboxProperties).

Step 3 populates the environment with the paths to src and temp, created in step 2,
while also populating OUT with the selected output path. It also puts a bunch of
Android stuff in the path if they were present in the Genrule constructor. While
it's probably worth untangling the Android stuff out of here at some point, this
commit makes no attempt to do so.

Reviewed By: cjhopman

shipit-source-id: 3a46098fb6