Ignore --exclude-incompatible-targets option
Summary:
This option is confusing because it is off by default:

```
build build --target-platforms=//platform:rule //foo/bar/...
```

will fail if there's at least one incompatible target. So this option should be on by default.

In fact, it is on in all mode files in one of our repositories.

We could add `--include-incompatible-targets` option, but it is not really needed, because there's not much can be done with incompatible targets:
* they cannot be built
* `buck query` shows "incorrect" information about incompatible targets (with selects resolved to something unreasonable or even failing)
* we have `buck uquery` which does not filter uncompatible targets already

Eventually we should remove this option.

Reviewed By: sbalabanov

shipit-source-id: 3a01aa0f02