Introduce LegacyProviderInfoCollectionImpl
Summary:
This is a ProviderInfoCollection specifically for rules that are yet to be migrated.
During compatibility they only be able to propagate an empty collection.

Changes coming up will make the default provider collection enforce the presence
of a DefaultInfo object, so this class is necessary for the legacy cases

Reviewed By: philipjameson

shipit-source-id: 55b7b684e4