Make various macro Args classes static
Summary: ModernBuildRule does not allow anything captured as part of the serialization process to be a non-static member class. Since these macros are legal inputs to genrule, they too must be made static in order to be serialized. Luckily, none of these classes reference their host class, so there are no repercussions to doing this.

Reviewed By: mykola-semko

shipit-source-id: b3b4894047