Retain all methods for inlined SAM interfaces in class ABIs - reland of D17448026
Summary:
This is a reland of D17448026, which was reverted. The fix for the revert is in D17539058.

The Kotlin compiler creates an extra class for a SAM interface. If the class is to be used in an inline function, then the class name contains `$sam$i`, otherwise it is just `$sam`. Therefore we are able to look for `$sam$i` and if it exists, we should retain all the methods in the class (there should be just one, plus the `<init>`).

Reviewed By: asp2insp

shipit-source-id: 8c60980c6a