Only create the root directory once when generating MBR steps
Summary: When generating setup steps for a Buildable, MBR issues recursive RmSteps for all output directories of the buildable. Since the root path is always considered an output path for the purposes of output path calculation, this causes MBR to issue a RmStep for the root output directory, followed immediately by a MakeCleanDirectory step, which deletes the output directory again. This commit doesn't issue the RmStep for the root output directory, since we're about to re-create it.

Reviewed By: cjhopman

shipit-source-id: 5ef43450c9