update sh_binary() to use exec so signals are handled correctly
Summary:
This fixes buck's `sh_binary()` template to exec the shell script with
`exec` rather than spawning it in a subshell, and letting the wrapper
script continue to run until the script exits.

Running the script though a parent wrapper process causes problems
because the wrappeor script does not forward signals to the child.  If
another program runs the `sh_binary()` and then later wants to kill it,
this currently doesn't work, as the signal only gets sent to Buck's
wrapper script, and never gets delivered to the actual shell script
being run.

Reviewed By: mykola-semko

shipit-source-id: 9003bf7f3d