Merge ParsePipeline and ConvertingPipeline
Summary:
Before this diff we have two levels of virtual functions:

`ParsePipeline` defines pure virtual functions which are implemented in `ConvertingPipeline` which adds more pure virtual functions which are implemented in `RawTargetNodePipeline` and `RawTargetNodeToTargetNodeParsePipeline`, and these classes also implement some methods from `ParsePipeline`.

This indirection makes code harder to read, it seems to be a fake abstraction which exists I believe for historical reasons.

This diff simply pushes whole `ParsePipeline` implementation down to `ConvertingPipeline`.

This diff is a preparation to implementation of `default_new_platform` attribute which requires heavy modifications of these classes.

Reviewed By: bobyangyf

shipit-source-id: 6124fe74ec