Split ParserTargetNodeFactory into ParserTargetNodeFromAttrMapFactory and ParserTargetNodeFromRawTargetNodeFactory
Summary:
Previously there was one interface `ParserTargetNodeFactory<T>` where `<T>` is practically either `Map<String, Object>` or `RawTargetNode`.

This interface is unused generalization, because all three implementations implement the interface for specific type `<T>` (map or node), and all fields and parameters also specify concrete type `<T>`.

The problem is as usual code navigation and refactoring: e. g. navigating to interface from the call site it's hard to understand which of three implementations is used, while in fact, for example, `ParserTargetNodeFactory<Map<String, Object>>` was implemented only by one class: it took me long time to realize that. It is more obvious now.

This work (together with D17649348 and D17649019) is done to merge regular and configuration node loading: currently there are two different pipelines for them, which is buggy and overcompliated and with code duplication. (`RawTargetNodeToTargetNodeParsePipeline` constructor is invoked twice in `PerBuildStateFactoryWithConfigurableAttributes.create` with very different arguments).

Reviewed By: sbalabanov

shipit-source-id: 60f3285e0c