Preserve main thread stack trace on internal error
Summary:
While investigating a race in cache invalidation which was manifested by an exception thrown from `DaemonicCellState`, I was found very inconvenient that error message in Buck stderr only contains stack trace of a thread which throws exception (thread where future callback is executed), but not a stack trace of the main thread which terminates the build.

Now if future results in exception which is not `HumanReadableException`, that exception is wrapped in a new exception with stack trace of the main thread. See the example below which makes it clear that build fails in target spec resolver (in one particular case, other places should likely be patched as well).

Reviewed By: sbalabanov

shipit-source-id: 82e3eaf550