Simplify cache invalidation on configuration changes
Summary:
Previously `DaemonicParserState` had two functions:

* `getOrCreateNodeCacheForConfigurationTargets` to create a cache view which records affected files for later invalidation of the whole cache
* `getOrCreateNodeCache` to create a cache without recording affected files

It may cause a bug: if configuration target is fetched first as a regular target (not a configuration of other target), and then fetched as configuration target, second fetch is cache hit, and is not recorded in `DaemonicParserState.configurationBuildFiles` field, thus cache is not invalidated when configuration is changed.

Now `DaemonicCacheViewWithTrackingBuildFiles` is merged with `DaemonicCacheView` and the information about whether a target is a configuration target or regular target is passed as a parameter of the function.

But this diff is actually done not to fix this bug, but to simplify the implementation, which should help further changes in this module (like merging regular and configuration target loading or fixing a race in target loading/target invalidation).

Keeping in mind that proper dependency tracking instead of whole cache invalidation is still on TODO list.

Reviewed By: sbalabanov

shipit-source-id: fb422aabdf