completely split RawTargetNodePipeline and RawTargetNodeToTargetNodeParsePipeline
Summary:
Copy-paste `ConvertingPipeline` code to both classes and remove `ConvertingPipeline`.

`ConvertingPipeline` became very abstract and it need to become more abstract to re-implement D17678313 (the diff was reverted because cache worked incorrectly, and cache behavior need to be changed in one of `ConvertingPipeline` subclasses).

This diff is a mechanical push-down and inline of trivial functions and removal of unused parameters.

Diff stat shows that the number of lines decreased by 40 (or stayed the same if we exclude imports of `ConvertingPipeline`), which kinda proves that `ConvertingPipeline` became wrong abstraction over time.

Reviewed By: sbalabanov

shipit-source-id: f52b788752