Fix bootclasspath related tests on Java 11
Summary:
The JVM system property `sun.boot.class.path` no longer exists in Java 9+. We used this property in tests to determine the system bootclasspath jar path for the currently executing JVM. We can't do this in Java 9+, even in principle, regardless of the behavior of this particular property, due to the introduction of the module system. The system bootclasspath isn't a jar (or even a set of jars) anymore.

The fix is to use the JDK stub jar we already have in the repo instead, which serves our needs just as well. We just need the path to a jar that defines all the system types that the compiler expects to see.

Reviewed By: stepancheg

shipit-source-id: d88b3eb7d4