Add a few more events to the machine log
Summary: The machine-readable log already outputs a few useful events, so this this commit has it output a few more. There's some room for cleanup in this class in follow-up diffs. With this commit, there's enough information to do some rudimentary graph width calculations and build an interval graph of the build tasks that Buck executed.

Reviewed By: rajyengi

shipit-source-id: c6a008ddb3