Rename ModuleMap to UmbrellaHeaderModuleMap, create ModuleMap interface
Summary:
Buck's module map support is currently limited to generating a module with an umbrella header. However, this requires libraries to actually have one. Buck has support for auto-generation, but it conflicts with the untracked header checker. Additionally, libraries with a header that has the same name as the library will not generate an umbrella header, since it would be a name conflict. This puts the burden on engineers to keep those headers valid umbrella headers.

Instead, we could generate a different form of module map that used either individual header declarations, using an implementation similar to the umbrella header generation, or just an "umbrella directory" declaration. However, we need to maintain the current behavior for open-source users. Therefore, this will be guarded behind a Buck config setting.

The first step is to rename the existing `ModuleMap` class to specify that it's for umbrella headers, and then to create a `ModuleMap` interface that can have alternative implementations. Then, we can fork behavior based on the Buck config setting.

Reviewed By: williamtwilson

shipit-source-id: 80227714e5