Allow creation of an "umbrella directory" module map
Summary:
This adds a separate mode for module map generation that uses an umbrella directory declaration, rather than an umbrella header declaration.

This fits Buck's general approach to headers better - it's kind of like an autoglob for every header in the library, without the need to either generate or manually maintain a valid umbrella header.

It's implemented by adding a new `ModuleMapMode` enum, and a new case to `HeaderMode` to differentiate between the two associated forms of module maps. I decided to have it specified at the `.buckconfig` level, rather than per-library, as this is simpler and is sufficient for the use-case.

Reviewed By: williamtwilson

shipit-source-id: 10842bbd24