Fix race in testMaterializeFiles
Summary:
OutputsMaterializerTest seems to break regularly on TravisCI when the WriteChannel for the 4th, oversized and thus stream API using artifact is not set.

This should fix this by closing the channel in the same place that batched fetches do, which we can do since we're not actually streaming in the test but handle all data in one request.

Reviewed By: rajyengi

shipit-source-id: 0c5678a1ca