Provide a more descriptive error message when Buck fails to parse the stdout for --show-output
Summary:
Adding `--show-output` to a `buck build` command in the command line displays output of the following format:
  <target name> <path relative to buck root>

E.g. For `buck build buck --show-output`
  //programs:buck buck-out/gen/programs/buck.pex

If there is no output, buck currently returns an error with a confusing message:
  "Expected: a collection with size <2>\n     but: collection size was <1>"

Modify the error message to make it more descriptive and reflective of the actual error.

Reviewed By: phyllipecesar

shipit-source-id: 012ec548e0