Make StringBuilder/String usage less confusing in thread rendering logic
Summary:
Current API is confusing because `MultiStateRenderer` takes a `StringBuilder` and returns a `String`. It's not clear whether a class implementing this interface should be appending to the builder or returning the `String` directly.

After looking into the history, it appears that the `StringBuilder` is intended for usage in `CommonThreadStateRenderer`, where it's forced to call `toString()` on the builder because `Ansi` only takes `String` params and not `StringBuilder` params. Additionally, `CommonThreadStateRender` (and `DistBuildSlaveStateRenderer`, but I guess this doesn't really matter since to my understanding it's getting deprecated) only ever adds two things maximum to the `StringBuilder` (lineWithoutStep and *maybe* the step), which means the effectiveness of StringBuilder is minimized. See https://dzone.com/articles/string-concatenation-performacne-improvement-in-ja, https://stackoverflow.com/questions/46512888/how-is-string-concatenation-implemented-in-java-9).

Considering these factors, I suggest removing `StringBuilder` from the API for clarity.

Reviewed By: rajyengi

shipit-source-id: 22402bb0d3