bzl files index
Summary:
This diff introduces the logic to maintain an inverse includes map  (an index from all files which are parse-time dependencies, including transitive ones, to packages which load them), which can be queried via `Index.getReverseIncludes()`.

Includes are stored as relative paths to a super root path (path of the topmost cell's path that roots all other cells )

This adds a new collection to `IndexGenerationData`: `IncludesMaps` that contains from two maps:
* `forwardMap` - from BUCK file to set of includes
* `reverseMap` - from include file to set of BUCK files

Like other generation maps(`ruleMap`, `buildPackageMap` and `rdepsMap`) it has its own lock.

Most of the new logic to keep `IncludesMaps` updated is in `determineDeltas()` that invokes `processIncludes()` from `Includes.kt`.

In the end, the code to implement `Index.getReverseIncludes()` is fairly simple
because it is primarily a lookup in the `includesMap.reverseMap`, the result of which are
build package paths that include passed file (ex. `.bzl` file).

Reviewed By: sbalabanov

shipit-source-id: a8fec563b6