Allow class names with spaces in AccumulateClassNamesStep
Summary:
kotlinc can spit out (extra, synthetic) classes that can have spaces in them, e.g. because the method that it is creating the class for had a space in.

Therefore we change the logic here from assuming two "parts" separated by a space, to just splitting on the last space (since the hash comes last and doesn't include a space).

Reviewed By: mykola-semko

shipit-source-id: 606505abaa