Fix buckd launch issue on Java 11
Summary: buckd fails to launch on a subset of machines on Java 11. This has something to do with the length and order of the full sequence of JVM args sometimes causing the JVM to ignore `-Xbootclasspath/a`. I don't have visibility into the JVM to figure out the true root cause and fully understand the circumstances under which this happens (and there's nothing in the documentation about this), but as the WARNING comment in the code says, making `-Xbootclasspath/a` the first arg seems to resolve the issue. Since this is affecting a bunch of users, I'd like to move forward with this fix ASAP.

Reviewed By: rajyengi

shipit-source-id: 18e8e85cae