Treating external tests
Summary: Before executing the tests, the idb option checks if there is an udid in the case of app and ui tests. This will not happen in external tests. So now we do this checking when the execution function is called, since the execution function is not called in the external tests.

Reviewed By: zeyadsalloum

shipit-source-id: c9428a8d06