Make Genrule a ModernBuildRule
Summary:
It's the same Genrule we know and love, but as a ModernBuildRule!

The previous diff in this stack copied the contents of Genrule to LegacyGenrule, so that this diff is free to convert Genrule (the parent class) to be a ModernBuildRule. This diff uses the previously-reviewed GenruleBuildable as the Buildable implementation.

The actual code changes in the Genrule class are relatively straightforward. Most of the code that previously generated steps or added to environment variables can be deleted outright since the functionality was moved to GenruleBuildable. The only things that remain are some getters and logic for constructing sandbox properties.

I had to fix a lot of integration tests in this diff. ModernBuildRule appends two underscores to the output directories of unflavored build targets, so any integration test that hardcoded a path to a genrule's output directory had to be updated. This was painful, but it's done.

One caveat about this diff is that it marks all Genrules as being unable to execute remotely. This is because I observed that it is often the case that genrules expect specific tools to be on the PATH that are not present in the remote environment. We'll want to investigate this further in follow-up tasks.

Reviewed By: mykola-semko

shipit-source-id: 5983cb4cb8