Extract SynlinkProxyArg to static inner class
Summary:
Like D17539453, D17557066, and D17579771, ModernBuildRule does not allow anything captured in the transitive closure of a buildable to be an anonyous class. This converts (yet another) anonymous class to a private static class so that it can be captured by ModernBuildRules.

I'd like to apply some static analysis to this so that we can stop playing whack-a-mole with random Buck crashes, but alas, for now fixing them as they pop up is effective.

Reviewed By: rajyengi

shipit-source-id: 8884b72f3b