Record GC collection events on the event bus
Summary: Adds GC collection events to the Buck event bus so that we can correlate Buck events with potentially large GCs. Any consumer of the Buck event bus can subscribe to these events; this commit amends the MachineReadableLoggerListener to dump them to the machine readable log, but any log consumer can subscribe to them. In a follow-up diff I will try to strap it to the Chrome trace viewer to see if we can get GCs to show up there.

Reviewed By: jtorkkola

shipit-source-id: b73177b53a