Fix label-based filtering
Summary:
Label based filtering (`--include <label>` and `--exclude <label>`) was broken due to querying the target graph for targets
after it had been expanded to include tests. This meant that the tests were always marked as "explicitly requested" so the label
filtering did not apply. This diff fixes the bug by saving the list of explicit targets prior to the test-target-graph-expansion.

Reviewed By: cute-jumper

shipit-source-id: eae979b2de