Remove new configurations CPU/OS detectors
Summary:
Remove `cpu_constraint_detector` and `host_constraint_detector` rules.

These rules are not used, not planned to be used soon, and it's better to remove them sonner, because:
* they get in the way of other changes
* to prevent someone from using them

These rules are easy to reimplement if they are needed, but I have a feeling we are going different direction with platform detection (detect `platform`, not individual constraints).

Reviewed By: bobyangyf

shipit-source-id: 254010d67c