Add blacklist query to android_app_modularity and android_binary
Summary:
Currently there exists a blacklist of targets which won't be pulled into any non-root voltron modules. In order to allow for greater flexibility we want to change this to a list of queries.
This query is being resolved in the same manner as [deps_query](https://buck.build/rule/android_library.html#deps_query).

Reviewed By: JamieEi

shipit-source-id: d0e592085c