Pass the original client PWD from wrapper to java
Summary: Currently we set the PWD when we execute buck to the project root. In order to support expanding relative build targets we need to get the original working directory, so pass that as a client env var. This makes consistent logic in java codebase easier between NG and nonNG runs.

Reviewed By: stepancheg

shipit-source-id: c53390ff86