Allow relative targets on the command line for most commands
Summary:
Update the CommnadLineTargetNodeSpecParser to relativize build targets provided on the command line against the PWD where the wrapper was invoked. This allows something like the following to work:

```
$ find . -type f
./BUCK
./foo/BUCK
./foo/bar/BUCK
$ cd foo
$ buck build bar: # Builds //foo/bar:
$ buck build ... # Builds //foo/...
$ buck build : # Builds //foo:
$ buck build bar:baz # Builds //foo/bar:baz
.....
```

Absolute targets still work, and this behavior can be disabled with a configuration flag. The only commands that do not use this path, and will be updated in a future diff, are perf commands, some audit commands, and the 'run' command.

Reviewed By: stepancheg

shipit-source-id: cbbc80ee23