Fix log uploading with Java 11 Buck with Java 8 on the PATH
Summary:
`UploaderLauncher` relies on Java on the path to create its asynchronous uploader child process. The Java code we execute will be Java 11 code in Java 11 builds of Buck, so this is problematic when Java 8 is on the `PATH`.

The fix is to use the same Java binary we're currently executing with for the child process. Note that the `java.home` system property always returns a value, irrespective of whether `JAVA_HOME` is set, though the latter will affect the former.

Reviewed By: rajyengi

shipit-source-id: 08f3734ff7