Introduce a helper to resolve Android SDK relative to project workspace
Summary:
Tests are often making build-tools version checks or using bits of the sdk based on defaults. This is problematic in integration tests that have an overridden `.buckconfig`, because it gets ignored. Additionally, the sdk checking code is duplicated in a couple of different places.

This introduces a helper that can be used to resolve sdk bits relative to a project workspace or filesystem, and uses it in one test. A subsequent diff will change the AssumeAndroidPlatform to use this helper.

Reviewed By: cute-jumper

shipit-source-id: 3ddd3cde1a