Test dependency stack is propagated when select() is not available
Summary:
Building a target `//:leaf` which depends on `//:intermediate` which depends on misconfigured `//:dep`.

Before this diff the error was:

```
Cannot use select() expression when target platform is not specified
```

without any clue what target needed selects, and how we got there from user requested target.

Now the error is:

```
Cannot use select() expression when target platform is not specified
    At //:dep
    At //:intermediate
    At //:leaf
```

Reviewed By: bobyangyf

shipit-source-id: 8d4e4e8987