Support early cut off computation for legacy compatibility
Summary:
We skip traversing the graph down if the current rule isn't compatible.
There's no point trying to get provider dependencies if the current rule
being requested is not going to use it.

This skips traversing the graph twice and avoids doubling the graph construction time
when compatibility mode is enabled

Reviewed By: philipjameson

shipit-source-id: 23cb31cfd2