Merge BuckQueryIntegrationTest and QueryCommandIntegrationTest
Summary:
`BuckQueryIntegrationTest` and `QueryCommandIntegrationTest` are about the same thing (moreover, the use the same test data), so merge tests in one class.

Probably we need to split the test because it's relatively large, but if we do, we should follow some logic, like put "deps" test in one class and "rdeps" tests in another.

Reviewed By: philipjameson

shipit-source-id: 199e64a204