rust: allow crate root to be specified for generated sources
Summary:
The `mapped_srcs` field allows a generated source to be mapped into
the crate's build container so that the source file can interact
with other sources in the normal Rust way. However this didn't work
for the root module (eg `lib.rs`/`main.rs`) because the generated
source didn't participate in the normal heuristic for identifying
the root module. The normal workaround would be to use the `crate_root`
attribute, but that was using `SourcePath`, and so failed at parse
time if the named file didn't exist.

This diff makes mapped_srcs participate in root module finding, and
changes `crate_root` to plain `String` so it can name paths which don't
exist at parse time.

Reviewed By: bobyangyf

shipit-source-id: c53512d5cc