Return getBuildTarget for AbiJar when not generating one
Summary:
This is a follow-up to the (hopefully temporary) hack that was added in D17861826.

There is an assumption that we will have an ABI jar of some sort available for all the deps that we need when building a source-only ABI, and so this new idea of not generating one at all breaks in that case. Instead, we can just use the full ABI jar in those cases (we don't have to generate it as it is prebuilt, so we're only paying the extra cost of loading it onto the classpath).

Reviewed By: asp2insp

shipit-source-id: dab49f7ff9